Home Page Displays Users
I just updated the home page. I replaced the placeholder text box thing that was previously on the home page. Now it should show a list of all of the users that have signed on as well as their IP addresses to make contacting them easier. I need to now pair this with a list of items so I can display that as well. The biggest challenge right now is trying to get a collection inside of a User document, that way each user can have their own collection of items.